Replace http module with Express, split routes into modules
Pure refactoring - no behavior changes. Moves all route logic out of
server.js into routes/api.js, routes/auth.js (placeholder), and
routes/static.js. server.js is now just app setup and listen.

Key translations from raw http to Express:
- Manual body accumulation (req.on data/end) -> express.json() + req.body
- URL.searchParams.get() -> req.query
- res.writeHead() + res.end(JSON.stringify()) -> res.status().json()
- fs.readFile static serving -> express.static()
- if/else routing -> express.Router()

Note: /api/why was never streaming - it uses client.messages.create()
and waits for the full response before replying. The Express version
preserves this behavior identically. Manually verified Why? button
works correctly after migration.

Also installs: express-session, passport, passport-google-oauth20, pg,
connect-pg-simple (unused here, needed for issues #91-#93).
