+++ routes/auth.js
@@ -1,0 +1,8 @@
+const express = require("express");
+
+const router = express.Router();
+
+// Auth routes will be added in issue #92 (Google OAuth with Passport).
+// This file is a placeholder to establish the module structure.
+
+module.exports = router;
